Split up prompt and search WIP
