+++ packages/react/package.json
@@ -31,7 +31,8 @@
     "p-debounce": "^4.0.0",
     "react-markdown": "^8.0.7",
     "react-merge-refs": "^2.0.2",
-    "remark-gfm": "^3.0.1"
+    "remark-gfm": "^3.0.1",
+    "type-fest": "^3.12.0"
   },
   "devDependencies": {
     "@testing-library/jest-dom": "^5.0.0",
