+++ packages/react/src/References.tsx
@@ -3,8 +3,8 @@ import { animated, useSpring } from '@react-spring/web';
 import React, { useCallback, useMemo, type ReactElement } from 'react';
 
 import { DEFAULT_MARKPROMPT_OPTIONS } from './constants.js';
-import { useMarkpromptContext } from './index.js';
 import * as Markprompt from './index.js';
+import { useMarkpromptContext } from './index.js';
 import { useElementSize } from './useElementSize.js';
 
 type ReferenceProps = {
@@ -73,7 +73,7 @@ const References = (props: ReferencesProps): ReactElement | null => {
     getLabel,
     transformReferenceId,
   } = props;
-  const { state, references } = useMarkpromptContext();
+  const { references, state } = useMarkpromptContext();
   const [ref, { height }] = useElementSize<HTMLDivElement>();
 
   const ReferenceComponent = useCallback(
